[dependency] - Add select eslint-plugin-jsx-a11y rules to lint config. (#175)
This project is a really good use case to apply some accessibility
rules to the lint config. For beginners getting started with React, we
can also teach/enforce basic accessibility rules in the web at author
time.

For now, I’ve just applied to the rules that are listed in
`eslint-config-airbnb` because these are most real-world tested, and we
can continue to add.

Happy to help if there are any questions/issues/concerns about the
plugin!